Add Simulation infrastructure and Ginzburg-Landau method.
This commit adds the generic infrastructure for running a gradient descent simulation, given a method of computing the energies of magnetization fields and a method of computing fixed points. This ability is encapsulated by the FixedPointMethod abstract class, which provides methods for computing the energies of magnetization fields and fixed points. We also implement the FixedPointMethod associated with the Ginzburg-Landau model, as described in Chapter 3 of "Topological solitons in two-dimensional chiral magnets".